Process some clip art multiple times in a sample.
Some clip art is simpler than others. For example, the MRIMachine is so simple that it took 2ms (previously). If we assume the margin of error for these sample sizes is approx 50-100ms: a 2ms sample size is pretty bad. So with this commit we process that clip art 100x in a sample, and now it takes around 163ms. This is more statistically meaningful information.

I considered just deleting the simple clip art, but I decided against it. Simple I want these tests to cover a wide range of possible usages: simple usages should be considered to. Instead I can scale up the test to better measure "simple clip art" and detect if we're making improvements or not.