Alt resolution ScaledMaskOutlineEngineTests#testReuse
With this commit the previous resolution is no longer strictly necessary. Now it functions as an optimization and not a bug fix.

This speaks to a broader problem with how the RectangleMasks interpret shapes: the outline can be be too big on the right and bottom edges. This can happen to any shape, but it's especially bad for rectangles since we rely on iterating over rectangles for certain operations.