Removing TandemIterator
I tried setting up a performance profile between the orig contains/intersects method and the TandemIterator implementation. The TandemIterator did show a subtle performance gain, but more importantly: it should inaccurate results sometimes.

Instead of fixing it (was it brittle enough to break again later?) I decided I'd rather delete it and stick to the safe (if slower) implementation for now.