Archiving an alt approach
This new approach attempts to reduce how often we create Areas by splitting apart AppendedShapes and only creating Areas for the absolutely essential parts. The hope is that there are enough free-floating shapes that we never NEED to convert to Areas, and if we can avoid that cost it'll improve overall efficiency.

... but in reality: it didn't work out. The new max % is 150%. (So with this option active: the TubmanEngine can take 150% of the time the Area class takes.) The average time is about the same.

However the "Gift Box" test case is the amazing success story here. It went from 97.907% to 7.667%. But it's the only test case that saw similar gains. IMO the risk is not worth the reward. It is probably worth backing up for posterity, though.