Increasing number of samples for comparison
Also (not shown in this commit): I tried another several approaches to optimize things. Mostly:

A. Multithreading. I was really surprised this didn't make an impact on performance. When we have to instantiate two Areas before calling area1.add(area2), I wrapped construction of one of those Areas in a second thread. It made such a small impact it might have been within the margin of error (around 1-2%).

B. I tried a new unique approach with a RectangleMask2D. The AddingShape currently goes through 3 tests for intersection/contains:
1. does the overall bounds intersect/contain the arg
2. does the member's bounds intersect/contain the arg
3. does the member itself intersect/contain the arg

I tried to add a new layer between 2 and 3: a RectangleMask2D. I've tried this multiple times before, and the general pattern is: it significantly helps one or two use cases, but is such an overall burden that it increases the max time to up to 150+% of the Area's original performance. (That is: it might REALLY help certain clip art uses cases, but it also really hurts others.) So with this approach I tried to be as lazy as possible: I created the mask using the member's bounds. I only subtracted from when I already HAD to consult shape#intersects(rect). If that call (shape#intersects(rect)) returned false: I knew I could subtract rect from the mask. I confirmed this was being invoked and we were chipping away the mask, but it didn't really impact performance. It also didn't *hurt* performance very much, but it seemed like overkill for minimal returns.

C. I tried another approach with masks: create a mask for each member's bounds and partition it. For example: split that mask up into a 3x3 grid, and check to see whether each cell is contained in the member shape (and subtract it from the mask if possible). I ended up testing NxN grids, where N=[1,20]. Most clip art tests followed a pretty clear curve with a trough (although some were spikier than others), but the low point of the trough mostly corresponded to the existing minimum. Having a clearly defined mask did not seem to help performance for than 1-2 percentage points.

All of these approaches were so underwhelming I ended up not keeping them.