Rejected: this uses a lazy mask
This appears to only reduce the average performance by about 1.1%, which could just be noise in the profiler. I'm leaving this in the git commit history, but I'll roll this change back.

(I referenced this change in a commit comment yesterday. I figured since I (slightly) optimized the RectangleMask2D I should try it again and save it for posterity.)