Issue #1: Adding OutlineOperation.Type.TRANSFORM
This includes a couple of new(ish) unit tests. I ran the EngineAccuracyTests up to 57,000,000 times overnight without a problem. This test is an adapation of a previous test somewhere in the git history.

Also the testMemoryFailure tests an OutOfMemoryError observed with the Area class with a particular sequence of steps. I think the OptimizedEngine avoids this error by luck (because it rearranges the order of certain operations) more than anything else.