Adding a couple of methods to help purge null-ops from the queue.
The first method (removeLeadingHiddenOperations) may be completely trivial and unlikely to present itself in the real world.

The second method (removeOperationsOutsideOfClipping) may help save some expensive Area#add invocations if they affect an
area that is outside our eventual clipping. This optimization may not be triggered depending on exactly how the Outline's bounds are flushed, though. (For ex: if a bunch of complex calls are made to Outline#add, and then Outline#getBounds2D is called: that forces a flush. Even if we later see a clipping and learn that it's OK to purge a lot of those added shapes.)