Running profiler results again
Also I'm removing CompoundShapeEngine from the profiler results. The results should as easy to understand as possible. (There's already a lot of explanation required in going through them...) The CompoundShapeEngine isn't something I currently want to talk about in a performance profile. We'll still use the CompoundShapeEngine for the accuracy tests, though, to make sure in our randomized tests it produces correct results.