Adding new failing unit test
The GeomCubicSolver fails because it doesn't handle near-degenerated cubics (yet), and the PolynomialCubicSolver also fails (rounding error somewhere?)

I was expecting the PolynomialCubicSolver to fail in general as we transitioned into a near-degenerated cubic (because the current threshold we impose to "fix" this is arbitrary), but this test doesn't capture that. I'll try to explore this more later.