Ran overnight clipping tests
This branch is currently focusing on the CubicSolver class, and currently the CubicSolver is only used for clipping.

This commit shows a modest (but visible) performance improvement with the current CubicSolver implementation.

Also noteworthy: the CubicSolverTests still include some failing unit tests. But all our other tests appear to pass. So while there are still some deficiencies in the CubicSolver classes: they don't appear to affect any other tests/usage so far.