Fixing more unit test failures
Previously I treated a cubic equation like it was a quadratic (when there was a wide disparity in the (x^3) and (x^2) coefficients) to safely find 1 root. Now instead I'm experimenting with what roots we can identify (machine error and all), then refining them, and then using synthetic division to try an alt path to identifying the other roots.