Failed experiment re double roots
There's a test (in the "basic" test that by default identifies 3 roots, when it should only identify 2. 2 of the 3 are very close to each other (something like -8.00000001 and -7.99999999).

In this case part of the problem is we found two x values for which f(x) = 0.

In this commit I theorized that there was a whole range of adjacent x-values that continued to produce f(x) = 0. So if I identified that two x values were part of the same range: I could reasonably guess it was a double root.

Unfortunately that's now how it works. In this case I saw results like:
f(-8.00000001) = 0
f(-8.000000005) = 1e-14
f(-8.000000001) = 0

(Those aren't literal examples, but you get the idea.) There wasn't a continuous run of x-values that produces a zero.

I'm committing this just to back it up in the history, but I'll roll this back shortly.