Adding failing unit test from largeDoubles
So... previously I thought this was passing. It turns out it was just throwing an exception, which triggered the fallback solver. So the only that changed here is:
A. We're highlighting the failure
B. It's failing in a subtly different way than before (instead of throwing an exception it's returning the wrong values)